store/requirements.txt
dj-database-url==0.5.0
Django==2.0.2
django-bootstrap4==0.0.6
django-heroku==0.3.1
djangorestframework==3.7.7
gunicorn==19.7.1
Pillow==5.0.0
psycopg2==2.7.4
pytz==2018.3
whitenoise==3.3.1
